begin integrating svg into new approach
