+++ frontend/index.html
@@ -6,6 +6,7 @@
 		<link rel="stylesheet" type="text/css" href="css/index.css">
 	</head>
 	<body>
+		<div id="svg-target"></div>
 		<main>
 		</main>
 	</body>
